refactor: Move non-binary root packages into common/
Move constants/, log/, telemetry/, and xds/ under common/ to consolidate
shared infrastructure alongside existing common packages (file, manager,
must, retry, spire). registry/ stays at root as it is domain logic.
