+++ agent/internal/cni/server/BUILD.bazel
@@ -18,9 +18,9 @@ go_library(
         "//agent/storage",
         "//agent/types",
         "//api/aether/cni/v1:cni",
-        "//constants",
+        "//common/constants",
+        "//common/xds",
         "//registry",
-        "//xds",
         "@build_buf_go_protovalidate//:protovalidate",
         "@com_github_go_logr_logr//:logr",
         "@com_github_grpc_ecosystem_go_grpc_middleware_v2//interceptors/protovalidate",
@@ -49,7 +49,7 @@ go_test(
         "//agent/types",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "//registry",
         "@com_github_envoyproxy_go_control_plane_envoy//admin/v3:admin",
         "@com_github_go_logr_logr//:logr",
@@ -86,7 +86,7 @@ go_test(
         "//agent/types",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "//registry",
         "@com_github_envoyproxy_go_control_plane_envoy//admin/v3:admin",
         "@com_github_go_logr_logr//:logr",
