+++ agent/internal/cni/server/pod.go
@@ -9,7 +9,7 @@ import (
 	"github.com/bpalermo/aether/agent/internal/xds/proxy"
 	"github.com/bpalermo/aether/agent/types"
 	cniv1 "github.com/bpalermo/aether/api/aether/cni/v1"
-	"github.com/bpalermo/aether/constants"
+	"github.com/bpalermo/aether/common/constants"
 	"github.com/bpalermo/aether/registry"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
