+++ agent/internal/xds/proxy/BUILD.bazel
@@ -20,7 +20,7 @@ go_library(
         "//agent/internal/xds/config",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "@com_github_envoyproxy_go_control_plane_envoy//config/cluster/v3:cluster",
         "@com_github_envoyproxy_go_control_plane_envoy//config/core/v3:core",
         "@com_github_envoyproxy_go_control_plane_envoy//config/endpoint/v3:endpoint",
@@ -60,7 +60,7 @@ go_test(
     deps = [
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "@com_github_envoyproxy_go_control_plane_envoy//config/cluster/v3:cluster",
         "@com_github_envoyproxy_go_control_plane_envoy//config/core/v3:core",
         "@com_github_envoyproxy_go_control_plane_envoy//config/route/v3:route",
