+++ agent/internal/xds/server/BUILD.bazel
@@ -10,8 +10,8 @@ go_library(
         "//agent/internal/xds/cache",
         "//agent/storage",
         "//api/aether/cni/v1:cni",
+        "//common/xds",
         "//registry",
-        "//xds",
         "@com_github_go_logr_logr//:logr",
     ],
 )
@@ -29,8 +29,8 @@ go_test(
         "//agent/storage",
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
+        "//common/xds",
         "//registry",
-        "//xds",
         "@com_github_envoyproxy_go_control_plane//pkg/cache/types",
         "@com_github_envoyproxy_go_control_plane//pkg/cache/v3:cache",
         "@com_github_envoyproxy_go_control_plane//pkg/resource/v3:resource",
@@ -62,7 +62,7 @@ go_test(
         "//api/aether/cni/v1:cni",
         "//api/aether/registry/v1:registry",
         "//registry",
-        "//xds",
+        "//common/xds",
         "@com_github_envoyproxy_go_control_plane//pkg/cache/types",
         "@com_github_envoyproxy_go_control_plane//pkg/cache/v3:cache",
         "@com_github_envoyproxy_go_control_plane//pkg/resource/v3:resource",
