+++ common/constants/cni.go
