+++ common/log/log_test.go
