+++ common/manager/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     importpath = "github.com/bpalermo/aether/common/manager",
     visibility = ["//visibility:public"],
     deps = [
-        "//log",
-        "//telemetry",
+        "//common/log",
+        "//common/telemetry",
         "@com_github_go_logr_logr//:logr",
         "@com_github_spf13_cobra//:cobra",
         "@io_k8s_sigs_controller_runtime//:controller-runtime",
