+++ common/telemetry/options.go
