+++ registry/internal/cloudmap/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     visibility = ["//registry:__subpackages__"],
     deps = [
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "@com_github_aws_aws_sdk_go_v2//aws",
         "@com_github_aws_aws_sdk_go_v2_service_servicediscovery//:servicediscovery",
         "@com_github_aws_aws_sdk_go_v2_service_servicediscovery//types",
