+++ registry/internal/k8s/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//registry:__subpackages__"],
     deps = [
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "@com_github_go_logr_logr//:logr",
         "@io_k8s_api//core/v1:core",
         "@io_k8s_sigs_controller_runtime//pkg/client",
@@ -20,7 +20,7 @@ go_test(
     embed = [":k8s"],
     deps = [
         "//api/aether/registry/v1:registry",
-        "//constants",
+        "//common/constants",
         "@com_github_go_logr_logr//:logr",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
